requirements.txt
torchvision
numpy
opencv-python